add detail function

+++ src/Controller/ProductController.php
@@ -5,6 +5,7 @@ namespace App\Controller;
 use App\Entity\Product;
 use App\Form\ProductType;
 use Doctrine\Persistence\ManagerRegistry;
+use http\Message;
 use phpDocumentor\Reflection\DocBlock\Tags\Reference\Reference;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\File\Exception\FileException;
@@ -12,6 +13,7 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\String\Slugger\SluggerInterface;
+use function Symfony\Config\Framework\Workflows\type;
 
 
 class ProductController extends AbstractController
@@ -35,21 +37,81 @@ class ProductController extends AbstractController
         return $this->render ('product/detail.html.twig',['product'=>$products]);
     }
 
-#[route ('/', name: 'product_delete')]
+#[route ('/product/delete/{id}', name: 'product_delete')]
     public function deleteAction(ManagerRegistry $doctrine, $id)
     {
         $em = $doctrine->getManager();
-        $product = $em->getRepository('App:Product')->find($id);
-        $em->remove($product);
-        $em->flush();
+        $product = $em->getRepository('App\Entity\Product')->find($id);
+        if (!is_null($product)) {
 
-        $this->addFlash(
-            'error',
-            'Product deleted'
-        );
 
-        return $this->redirectToRoute('product_list');
+
+            $em->remove($product);
+            $em->flush();
+
+            $this->addFlash(
+                'error',
+                'Product deleted'
+            );
+        } else {
+            $this->addFlash(
+                'error',
+                'product not existed!'
+            );
+        }
+        return $this->redirectToRoute('app_product');
     }
+
+    #route ('/product/create',name: 'product_create', methods={"GET", "POST"})
+    public function createAction(ManagerRegistry$doctrine,Request $request, SluggerInterface $slugger)
+    {
+        $product = new Product();
+        $form = $this->createForm(ProductType::class, $product);
+
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            // uplpad file
+            $productImage = $form->get('productImage')->getData();
+            if ($productImage) {
+                $originalFilename = pathinfo($productImage->getClientOriginalName(), PATHINFO_FILENAME);
+                // this is needed to safely include the file name as part of the URL
+                $safeFilename = $slugger->slug($originalFilename);
+                $newFilename = $safeFilename . '-' . uniqid() . '.' . $productImage->guessExtension();
+
+                // Move the file to the directory where brochures are stored
+                try {
+                    $productImage->move(
+                        $this->getParameter('productImages_directory'),
+                        $newFilename
+                    );
+                } catch (FileException $e) {
+                    $this->addFlash(
+                        'error',
+                        'Cannot upload'
+                    );// ... handle exception if something happens during file upload
+                }
+                $product->setProductImage($newFilename);
+            }else{
+                $this->addFlash(
+                    'error',
+                    'Cannot upload'
+                );// ... handle exception if something happens during file upload
+            }
+            $em = $doctrine->getManager();
+            $em->persist($product);
+            $em->flush();
+
+            $this->addFlash(
+                'notice',
+                'Product Added'
+            );
+            return $this->redirectToRoute('app_product');
+        }
+        return $this->renderForm('product/create.html.twig', ['form' => $form,]);
+    }
+
+
 
     #[route('/', name: 'product_edit')]
     public function editAction(ManagerRegistry $doctrine, int $id, Request $request): Response
